Extracted all non-cookie-specific stuff from CookiesFieldEditor
